scoring for a game

+++ app/src/main/java/com/radlance/languageapp/domain/main/UserScore.kt
@@ -4,5 +4,5 @@ import com.radlance.languageapp.domain.auth.User
 
 data class UserScore(
     val user: User,
-    val score: Int
+    val score: Double
 )
